requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.12.3
black==24.4.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
comm==0.2.2
cryptography==42.0.5
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.0.4
django-allauth==0.61.1
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.8
fastjsonschema==2.21.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
isort==5.13.2
jedi==0.19.2
jinja2==3.1.5
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy-extensions==1.0.0
narwhals==1.26.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.2
oauthlib==3.2.2
openai==1.61.1
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.2.0
playwright==1.43.0
plotly==6.0.0
prompt_toolkit==3.0.50
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==11.1.0
Pygments==2.19.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python3-openid==3.2.0
pytz==2025.1
pyyaml==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.9.4
rich-toolkit==0.13.2
rpds-py==0.22.3
ruff==0.9.6
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.38
sqlmodel==0.0.22
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.45.3
tomli==2.0.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.1
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
